# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CinemaCreate:
    properties:
      actors:
        items:
          type: string
        type: array
      age_rating:
        type: integer
      description:
        type: string
      duration_minutes:
        type: integer
      genres:
        items:
          type: string
        type: array
      name:
        type: string
      original_title:
        type: string
      poster_url:
        type: string
      rating:
        type: number
      user_rating:
        type: integer
      year:
        type: integer
    required:
    - name
    type: object
  dto.CinemaCreateResponse:
    properties:
      id:
        type: string
    type: object
  dto.CinemaUpdate:
    properties:
      actors:
        items:
          type: string
        type: array
      age_rating:
        type: integer
      description:
        type: string
      duration_minutes:
        type: integer
      genres:
        items:
          type: string
        type: array
      name:
        type: string
      original_title:
        type: string
      poster_url:
        type: string
      rating:
        type: number
      year:
        type: integer
    type: object
  dto.CinemaView:
    properties:
      actors:
        items:
          type: string
        type: array
      age_rating:
        type: integer
      description:
        type: string
      duration_minutes:
        type: integer
      genres:
        items:
          type: string
        type: array
      id:
        type: string
      name:
        type: string
      poster_url:
        type: string
      rating:
        type: number
      user_rating:
        type: integer
      year:
        type: integer
    type: object
...
  dto.HttpErr:
    properties:
      description:
        example: verbose error description
        type: string
      err:
        example: 'Short error message : ''Not Found'' | ''Internal Server Error''
          | etc'
        type: string
    required:
    - description
    - err
    type: object
  dto.OkResponse:
    properties:
      status:
        example: ok
        type: string
    required:
    - status
    type: object
...
host: https://prod-team-32-n26k57br.REDACTED
info:
  contact: {}
  title: MetaCinema
  version: "1.0"
paths:
  /admin/films:
    post:
      consumes:
      - application/json
      description: Создает фильм
      parameters:
      - description: access token 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request Body
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/dto.CinemaUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      security:
      - Bearer: []
      summary: create film
      tags:
      - admin
  /admin/films/{id}:
    delete:
      description: Удаляет фильм
      parameters:
      - description: access token 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of film
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      security:
      - Bearer: []
      summary: delete film
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Обновляет фильм
      parameters:
      - description: access token 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of film
        in: path
        name: id
        required: true
        type: string
      - description: Request Body
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/dto.CinemaUpdate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CinemaUpdate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      security:
      - Bearer: []
      summary: update film
      tags:
      - admin
...
  /films/:
    post:
      consumes:
      - application/json
      description: Добавляет новый фильм в подборку пользователя или добавляет глобальный
        фильм в базу админом
      parameters:
      - description: access token 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request Body
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/dto.CinemaCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CinemaCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      security:
      - Bearer: []
      summary: upload new film
      tags:
      - films
...
  /films/{FilmID}/picture:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: access token 'Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: UUID of the film
        in: path
        name: FilmID
        required: true
        type: string
      - description: File of the pic
        in: formData
        name: uploadfile
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      security:
      - Bearer: []
      summary: Set picture of the cinema
      tags:
      - films
  /films/{filmId}:
    get:
      consumes:
      - application/json
      description: Получить фильм по id
      parameters:
      - description: film uuid
        in: path
        name: filmId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CinemaView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: get film
      tags:
      - films
...
  /films/feed:
    get:
      consumes:
      - application/json
      description: Список рекомендованных фильмов
      parameters:
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: usernames
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CinemaView'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: get recommended films
      tags:
      - films
  /films/genres:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: Get genres
      tags:
      - films
  /films/popular:
    get:
      consumes:
      - application/json
      description: Получить список Самых Высоко оценённых фильмов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CinemaView'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: get top rated films
      tags:
      - films
  /films/search:
    post:
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.HttpErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.HttpErr'
      summary: search film
      tags:
      - cinemas
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"